+++ app/Jobs/DirtRally/ImportEventJob.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Jobs;
+namespace App\Jobs\DirtRally;
 
+use App\Jobs\Job;
 use App\Models\DirtRally\Event;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Queue\InteractsWithQueue;
